backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
python-dotenv==1.0.0
aiogram==3.10.0
pandas==2.1.3
openpyxl==3.1.2
aiohttp==3.9.1